started adding new config structure
