+++ Configs/system.ini
@@ -1,0 +1,23 @@
+[DEFAULT_VALUES]
+image_path = Images/Examples/example.png
+out_format = .blend
+overwrite_data = False
+blender_installation_path = C:\Program Files\Blender Foundation\Blender 2.93\blender.exe
+file_structure = [[[0,0,0], [0,0,0], [0,0,0]], [[0,0,0], [0,0,0], [0,0,0]], [[0,0,0], [0,0,0], [0,0,0]]]
+mode = CYLINDER
+
+[FEATURES]
+floors = True
+rooms = True
+walls = True
+doors = True
+windows = True
+
+[EXTRA_SETTINGS]
+remove_noise = True
+rescale_image = True
+
+[WALL_CALIBRATION]
+calibration_image_path = Images/Calibrations/wallcalibration.png
+wall_size_calibration = 33.5
+
