+++ Docker/docker-entrypoint.sh
@@ -14,7 +14,7 @@ elif [ "$1" = "jupyter" ]; then
     jupyter notebook /home/floorplan_to_blender/ --port=8888 --ip=0.0.0.0 --allow-root --no-browser --NotebookApp.token=''
 else
     echo "Starting ftb script"
-    python3 /home/floorplan_to_blender/create_blender_project_from_floorplan.py
+    python3 /home/floorplan_to_blender/main.py
 fi
 
 echo "--- Done ---"
